Move tests from test_optimize to test_history
